plagiarism-detector-back/requirements.txt
flask
torch
torchvision
numpy
pillow
PyPDF2
python-docx
PyMuPDF
